feat:下沉SematicBuffer至MemoryLibrary模块
